This change adds two things:   * Add  void *_objRef field to ApiMsg, defaulting to NULL   * Add a new method addEndpointObj to RestAPIEndpoints The new method is the same as the original addEndpointObj except that it takes an objRef parameter.  The library doesn't do anything with this parameter except that it gets passed back with the RestAPIEndpointMsg when the callback is called.
The purpose of this is to allwo you to use object (instance) methods
as callbacks by adding a simple little wrapper.  In my case, I make the
wrapper a private static

private:
	void handlePost(RestAPIEndpointMsg &msg, String &retStr);
	static void handlePostWrapper(RestAPIEndpointMsg &msg,
		String &retStr);

the wrapper is a static method, which can be passed in as a function
pointer.  The wrapper is short and sweet:

void
GrafanaListener::handlePostWrapper(RestAPIEndpointMsg& apiMsg, String& retStr) {
    GrafanaListener *that = (GrafanaListener *) apiMsg._objRef;
    that->handlePost(apiMsg, retStr);

}

What all of this does is lets your callbacks reference other methods and
instance members of the enveloping class, approaching something more
a bit more object-oriented.